requirements.txt
PySide6>=6.5
Flask>=2.3
Flask-Cors>=4.0
fastapi>=0.110
uvicorn>=0.23
requests>=2.31
websockets>=11.0
pandas>=2.0
openpyxl>=3.1
reportlab>=3.6
opencv-python>=4.8
pyzbar>=0.1.9
qrcode[pil]>=7.4
PyJWT>=2.8
boto3>=1.34
cryptography>=42.0
